+++ release_notes.md
@@ -13,6 +13,7 @@
 - APP-3748 - [Inputs] Add result Limit on resolving magic variables
 - APP-3747 - [API] Added support for Group -> Victim Asset Associations
 - APP-3754 - [API] Update group_model to address openapi spec discrepancy
+- APP-3756 - [KeyValue] Add KeyValueMock, an implementation of KeyValueABC only for testing and running apps locally.
 
 ### 3.0.3